ref(api): establish transport boundary for API client
Braintrust HTTP mechanics currently live inside logger.py, coupling transport to logging and leaving no isolated foundation for a resource-oriented API client. Establish this boundary first so retries, structured errors, and endpoint services can be centralized in follow-up work without mixing those behavior changes into the initial refactor.
