+++ py/src/braintrust/api/__init__.py
@@ -1,0 +1,3 @@
+"""Braintrust API client package."""
+
+__all__: list[str] = []
